Move lua methods to celx_celestia and celx_misc
